// server/src/ts/interface/wishlist.d.ts
export interface DomainSave {
  uid: string;
  bookId: string;
}

export interface DomainGet {
  uid: string;
  bookId: string;
}

export interface DomainDelete {
  uid: string;
  bookId: string;
}

export interface WebSave {
  uid: string;
  bookId: string;
}

export interface WebDelete {
  bookId: string;
}